feat: centralize tag helpers
